refactor: move constants file to utils directory
