+++ app/note/[eventid]/page.tsx
@@ -2,7 +2,7 @@
 import Note from "./Note";
 
 import { NostrProvider } from "nostr-react";
-import { PROFILE_RELAYS } from "../../constants";
+import { PROFILE_RELAYS } from "../../utils/constants";
 import { usePathname } from "next/navigation";
 import { useEffect, useState } from "react";
 
@@ -15,7 +15,8 @@ export default function NotePage() {
       setEventId(pathname.split("/").pop() || "");
       console.log("eventId from path name", eventId);
     }
-  }, []);
+  }, [pathname, eventId]);
+
   return (
     <>
       <NostrProvider relayUrls={PROFILE_RELAYS} debug={true}>
