+++ app/page.jsx
@@ -2,7 +2,7 @@
 import NoteArea from "./NoteArea";
 
 import { NostrProvider } from "nostr-react";
-import { RELAYS } from "./constants";
+import { RELAYS } from "./utils/constants";
 
 export default function HomePage() {
   return (
